// src/payload/payload-types.ts
export interface Page {
  id: string;
  title: string;
  publishedAt?: string;
  hero: {
    type: 'none' | 'highImpact' | 'mediumImpact' | 'lowImpact';
    richText: {
      [k: string]: unknown;
    }[];
    links?: {
      link: {
        type?: 'reference' | 'custom';
        newTab?: boolean;
        reference: {
          relationTo: 'pages';
          value: string | Page;
        };
        url: string;
        label: string;
        appearance?: 'default' | 'primary' | 'secondary';
      };
      id?: string;
    }[];
    media: string | Media;
  };
  layout: (
    | {
        invertBackground?: boolean;
        richText: {
          [k: string]: unknown;
        }[];
        links?: {
          link: {
            type?: 'reference' | 'custom';
            newTab?: boolean;
            reference: {
              relationTo: 'pages';
              value: string | Page;
            };
            url: string;
            label: string;
            appearance?: 'primary' | 'secondary';
          };
          id?: string;
        }[];
        id?: string;
        blockName?: string;
        blockType: 'cta';
      }
    | {
        invertBackground?: boolean;
        columns?: {
          size?: 'oneThird' | 'half' | 'twoThirds' | 'full';
          richText: {
            [k: string]: unknown;
          }[];
          enableLink?: boolean;
          link?: {
            type?: 'reference' | 'custom';
            newTab?: boolean;
            reference: {
              relationTo: 'pages';
              value: string | Page;
            };
            url: string;
            label: string;
            appearance?: 'default' | 'primary' | 'secondary';
          };
          id?: string;
        }[];
        id?: string;
        blockName?: string;
        blockType: 'content';
      }
    | {
        invertBackground?: boolean;
        position?: 'default' | 'fullscreen';
        media: string | Media;
        id?: string;
        blockName?: string;
        blockType: 'mediaBlock';
      }
    | {
        introContent: {
          [k: string]: unknown;
        }[];
        populateBy?: 'collection' | 'selection';
        relationTo?: 'posts' | 'projects';
        categories?: string[] | Category[];
        limit?: number;
        selectedDocs?:
          | (
              | {
                  relationTo: 'posts';
                  value: string;
                }
              | {
                  relationTo: 'projects';
                  value: string;
                }
            )[]
          | (
              | {
                  relationTo: 'posts';
                  value: Post;
                }
              | {
                  relationTo: 'projects';
                  value: Project;
                }
            )[];
        populatedDocs?:
          | (
              | {
                  relationTo: 'posts';
                  value: string;
                }
              | {
                  relationTo: 'projects';
                  value: string;
                }
            )[]
          | (
              | {
                  relationTo: 'posts';
                  value: Post;
                }
              | {
                  relationTo: 'projects';
                  value: Project;
                }
            )[];
        populatedDocsTotal?: number;
        id?: string;
        blockName?: string;
        blockType: 'archive';
      }
    | {
        profiles?: {
          size?: 'oneThird' | 'half' | 'twoThirds' | 'full';
          name: string;
          title: string;
          description: {
            [k: string]: unknown;
          }[];
          picture: string | Media;
          enableLink?: boolean;
          link?: string;
          id?: string;
        }[];
        id?: string;
        blockName?: string;
        blockType: 'profileEntries';
      }
  )[];
  slug?: string;
  meta?: {
    title?: string;
    description?: string;
    image?: string | Media;
  };
  updatedAt: string;
  createdAt: string;
  _status?: 'draft' | 'published';
}
export interface Media {
  id: string;
  alt: string;
  caption?: {
    [k: string]: unknown;
  }[];
  updatedAt: string;
  createdAt: string;
  url?: string;
  filename?: string;
  mimeType?: string;
  filesize?: number;
  width?: number;
  height?: number;
}
export interface Category {
  id: string;
  title?: string;
  parent?: string | Category;
  breadcrumbs?: {
    doc?: string | Category;
    url?: string;
    label?: string;
    id?: string;
  }[];
  updatedAt: string;
  createdAt: string;
}
// ...
export interface Project {
  id: string;
  title: string;
  categories?: string[] | Category[];
  publishedAt?: string;
  hero: {
    type: 'none' | 'highImpact' | 'mediumImpact' | 'lowImpact';
    richText: {
      [k: string]: unknown;
    }[];
    links?: {
      link: {
        type?: 'reference' | 'custom';
        newTab?: boolean;
        reference: {
          relationTo: 'pages';
          value: string | Page;
        };
        url: string;
        label: string;
        appearance?: 'default' | 'primary' | 'secondary';
      };
      id?: string;
    }[];
    media: string | Media;
  };
  layout: (
    | {
        invertBackground?: boolean;
        richText: {
          [k: string]: unknown;
        }[];
        links?: {
          link: {
            type?: 'reference' | 'custom';
            newTab?: boolean;
            reference: {
              relationTo: 'pages';
              value: string | Page;
            };
            url: string;
            label: string;
            appearance?: 'primary' | 'secondary';
          };
          id?: string;
        }[];
        id?: string;
        blockName?: string;
        blockType: 'cta';
      }
    | {
        invertBackground?: boolean;
        columns?: {
          size?: 'oneThird' | 'half' | 'twoThirds' | 'full';
          richText: {
            [k: string]: unknown;
          }[];
          enableLink?: boolean;
          link?: {
            type?: 'reference' | 'custom';
            newTab?: boolean;
            reference: {
              relationTo: 'pages';
              value: string | Page;
            };
            url: string;
            label: string;
            appearance?: 'default' | 'primary' | 'secondary';
          };
          id?: string;
        }[];
        id?: string;
        blockName?: string;
        blockType: 'content';
      }
    | {
        invertBackground?: boolean;
        position?: 'default' | 'fullscreen';
        media: string | Media;
        id?: string;
        blockName?: string;
        blockType: 'mediaBlock';
      }
    | {
        introContent: {
          [k: string]: unknown;
        }[];
        populateBy?: 'collection' | 'selection';
        relationTo?: 'posts' | 'projects';
        categories?: string[] | Category[];
        limit?: number;
        selectedDocs?:
          | (
              | {
                  relationTo: 'posts';
                  value: string;
                }
              | {
                  relationTo: 'projects';
                  value: string;
                }
            )[]
          | (
              | {
                  relationTo: 'posts';
                  value: Post;
                }
              | {
                  relationTo: 'projects';
                  value: Project;
                }
            )[];
        populatedDocs?:
          | (
              | {
                  relationTo: 'posts';
                  value: string;
                }
              | {
                  relationTo: 'projects';
                  value: string;
                }
            )[]
          | (
              | {
                  relationTo: 'posts';
                  value: Post;
                }
              | {
                  relationTo: 'projects';
                  value: Project;
                }
            )[];
        populatedDocsTotal?: number;
        id?: string;
        blockName?: string;
        blockType: 'archive';
      }
  )[];
  relatedProjects?: string[] | Project[];
  slug?: string;
  meta?: {
    title?: string;
    description?: string;
    image?: string | Media;
  };
  updatedAt: string;
  createdAt: string;
  _status?: 'draft' | 'published';
}